Stage and commit moved test files and assets to restore test discovery
