+++ my-app/package.json
@@ -24,9 +24,6 @@
     "fast-csv": "^5.0.2",
     "file-saver": "^2.0.5",
     "firebase": "^10.14.1",
-    "firebase-admin": "^12.7.0",
-    "firebase-functions": "^6.3.2",
-    "firebase-tools": "^14.0.1",
     "font-awesome": "^4.7.0",
     "js-cookie": "^3.0.5",
     "jszip": "^3.10.1",
@@ -34,7 +31,6 @@
     "leaflet.awesome-markers": "^2.0.5",
     "lucide-react": "^0.453.0",
     "luxon": "^3.6.1",
-    "nodemailer": "^6.10.0",
     "papaparse": "^5.5.2",
     "react": "^18.3.1",
     "react-big-calendar": "^1.15.0",
@@ -44,7 +40,7 @@
     "react-google-autocomplete": "^2.7.4",
     "react-icons": "^5.5.0",
     "react-router-dom": "^6.26.2",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
