+++ my-app/src/__tests__/minimal-with-test-utils.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '../setupTests';
-import { WithMuiTheme } from '../test-utils/withMuiTheme';
+import { WithMuiTheme } from '../../test-utils/withMuiTheme';
 
 describe('Minimal Test with withMuiTheme', () => {
   it('renders a div using WithMuiTheme', () => {
