+++ my-app/tests/LoadingIndicator.test.tsx
